requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
ta-lib==0.4.28
scipy==1.11.1
pytz==2023.3
logging-config==0.5.0
apscheduler==3.10.4
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0